% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_plotly_value_text}
\alias{add_plotly_value_text}
\title{Add Plotly Value Text}
\usage{
add_plotly_value_text(tbl, cols)
}
\arguments{
\item{tbl}{A tibble with column text}

\item{cols}{A vector of strings that are columns in the tibble}
}
\description{
Add Plotly Value Text
}
