% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot_ui.R
\name{barplot_ui}
\alias{barplot_ui}
\title{Barplot UI}
\usage{
barplot_ui(
  id,
  title = "",
  barchart_html = htmltools::includeMarkdown(get_markdown_path("barchart1")),
  ...
)
}
\arguments{
\item{id}{Module ID}

\item{title}{A string}

\item{barchart_html}{A string that is HTML}

\item{...}{Arguments passed to drilldown_scatterplot_ui}
}
\description{
Barplot UI
}
