% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements_ui.R
\name{categorical_model_covariate_element_ui}
\alias{categorical_model_covariate_element_ui}
\title{Categorical Model Covariate Element UI}
\usage{
categorical_model_covariate_element_ui(id)
}
\arguments{
\item{id}{A shiny ID}
}
\description{
Categorical Model Covariate Element UI
}
