% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_object_functions.R
\name{cohort_has_classes}
\alias{cohort_has_classes}
\title{Cohort Has Classes}
\usage{
cohort_has_classes(cohort_object, classes, all_classes = T)
}
\arguments{
\item{cohort_object}{A named list}

\item{classes}{A vector of strings}

\item{all_classes}{A logical}
}
\description{
Cohort Has Classes
}
