% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_object_functions.R
\name{cohort_has_dataset}
\alias{cohort_has_dataset}
\title{Cohort Has Dataset}
\usage{
cohort_has_dataset(cohort_object, datasets)
}
\arguments{
\item{cohort_object}{A named list}

\item{datasets}{A vector of strings}
}
\description{
Cohort Has Dataset
}
