% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_object_functions.R
\name{cohort_has_features}
\alias{cohort_has_features}
\title{Cohort Has Features}
\usage{
cohort_has_features(cohort_object, features, all_features = T)
}
\arguments{
\item{cohort_object}{A named list}

\item{features}{A vector of strings}

\item{all_features}{A logical}
}
\description{
Cohort Has Features
}
