% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_object_functions.R
\name{cohort_has_group}
\alias{cohort_has_group}
\title{Cohort Has Group}
\usage{
cohort_has_group(cohort_object, groups)
}
\arguments{
\item{cohort_object}{A named list}

\item{groups}{A vector of strings}
}
\description{
Cohort Has Group
}
