% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_selection_server.R
\name{cohort_selection_server}
\alias{cohort_selection_server}
\title{Cohort Selection Server}
\usage{
cohort_selection_server(id, ...)
}
\arguments{
\item{id}{A shiny ID}

\item{...}{unused args}
}
\description{
Cohort Selection Server
}
