% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_dataset_selection_functions.R
\name{create_cohort_module_string}
\alias{create_cohort_module_string}
\title{Create Cohort Module String}
\usage{
create_cohort_module_string(.datasets, tbl = NULL)
}
\arguments{
\item{.datasets}{a string(s) in the dataset column of the tibble}

\item{tbl}{A tibble with columns dataset, and module}
}
\description{
Create Cohort Module String
}
