% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_nested_named_list}
\alias{create_nested_named_list}
\title{Create Nested Named List}
\usage{
create_nested_named_list(
  tbl,
  names_col1 = "feature_class",
  names_col2 = "feature_display",
  values_col = "feature_name"
)
}
\arguments{
\item{tbl}{A tibble with the below columns}

\item{names_col1}{A column that will be the names of the top list}

\item{names_col2}{A column that will be the names of the nested lists}

\item{values_col}{A column that will be the values of the nested lists}
}
\description{
Create Nested Named List
}
