% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_plotly_text}
\alias{create_plotly_text}
\title{Create Plotly Text}
\usage{
create_plotly_text(tbl, name, group, cols, title)
}
\arguments{
\item{tbl}{A tibble}

\item{name}{A column}

\item{group}{A column}

\item{cols}{A vector of strings, which are columns in the tibble}

\item{title}{A string}
}
\description{
Create Plotly Text
}
