% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_table_server.R
\name{data_table_server}
\alias{data_table_server}
\title{Data Table Server}
\usage{
data_table_server(
  id,
  tbl,
  options = list(pageLength = 10),
  color = F,
  color_column = NULL,
  colors = NULL,
  ...
)
}
\arguments{
\item{id}{a shiny id}

\item{tbl}{at tibble}

\item{options}{sent to DT::datatable}

\item{color}{T/F}

\item{color_column}{a column in the tbl}

\item{colors}{sent to DT::styleEqual}

\item{...}{args sent to DT::datatable}
}
\description{
Data Table Server
}
