% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_table_ui.R
\name{data_table_ui}
\alias{data_table_ui}
\title{Data Table UI}
\usage{
data_table_ui(id, message_html = "", button_text = "Download datatable")
}
\arguments{
\item{id}{a shiny id}

\item{message_html}{sent to messageBox}

\item{button_text}{a string}
}
\description{
Data Table UI
}
