% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions_plot_server.R
\name{distributions_plot_server}
\alias{distributions_plot_server}
\title{Barplot Server}
\usage{
distributions_plot_server(
  id,
  plot_data_function,
  features = shiny::reactive(NULL),
  distplot_xlab = shiny::reactive(""),
  scale_method_default = shiny::reactive("None"),
  feature_default = shiny::reactive(NULL),
  drilldown = shiny::reactive(F),
  ...
)
}
\arguments{
\item{id}{Module ID}

\item{plot_data_function}{A shiny::reactive that returns a function
The function must take an argument called ".feature" and return a
dataframe with columns "sample", "feature", "feature_value", "group",
and optionally "group_description", "color"}

\item{features}{A shiny::reactive that returns a dataframe with "feature",
"feature_display", and any other additional optional columns to group the
features by}

\item{distplot_xlab}{A shiny::reactive that returns a string}

\item{scale_method_default}{A shiny::reactive that returns a string}

\item{feature_default}{A shiny::reactive that returns a string}

\item{drilldown}{A shiny::reactive that returns True or False}

\item{...}{shiny::reactives passed to drilldown_histogram_server}
}
\description{
Barplot Server
}
