% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions_plot_ui.R
\name{distributions_plot_ui}
\alias{distributions_plot_ui}
\title{Distributions Plot UI}
\usage{
distributions_plot_ui(id, title = "")
}
\arguments{
\item{id}{Module ID}

\item{title}{A string}
}
\description{
Distributions Plot UI
}
