% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drilldown_histogram_server.R
\name{drilldown_histogram_server}
\alias{drilldown_histogram_server}
\title{Drilldown Histogram Server}
\usage{
drilldown_histogram_server(id, plot_data, eventdata, ...)
}
\arguments{
\item{id}{Module ID}

\item{plot_data}{A shiny::reactive that returns a dataframe with columns
"group", "feature_value"}

\item{eventdata}{A shiny::reactive that returns a dataframe with column
"key"}

\item{...}{arguments sents to plotly_histogram}
}
\description{
Drilldown Histogram Server
}
