% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drilldown_scatterplot_server.R
\name{drilldown_scatterplot_server}
\alias{drilldown_scatterplot_server}
\title{Drilldown Scatterplot Server}
\usage{
drilldown_scatterplot_server(id, plot_data, eventdata)
}
\arguments{
\item{id}{Module ID}

\item{plot_data}{A shiny::reactive that returns a dataframe with columns
"sample", "group", "feature", "feature_value"}

\item{eventdata}{A shiny::reactive that returns a dataframe with column
"key"}
}
\description{
Drilldown Scatterplot Server
}
