% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_object_functions.R
\name{get_cohort_feature_class_list}
\alias{get_cohort_feature_class_list}
\title{Get Cohort Feature Class List}
\usage{
get_cohort_feature_class_list(cohort_object)
}
\arguments{
\item{cohort_object}{A named list}
}
\description{
Get Cohort Feature Class List
}
