% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_system_path_file}
\alias{get_system_path_file}
\title{Get System Path File}
\usage{
get_system_path_file(prefix, extension, folder, package = "iatlas.modules2")
}
\arguments{
\item{prefix}{A string, the file prefix}

\item{extension}{A the string, the file extension}

\item{folder}{A string, the file's folder}

\item{package}{A string, the package the file is in}
}
\description{
Get System Path File
}
