% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements_server.R
\name{group_filter_element_server}
\alias{group_filter_element_server}
\title{Group Filter Element Server}
\usage{
group_filter_element_server(
  id,
  reactive_values,
  module_id,
  group_named_list,
  datasets
)
}
\arguments{
\item{id}{A shiny ID}

\item{reactive_values}{a shiny::reactiveValues() object}

\item{module_id}{A unique value for this instance of the element being called}

\item{group_named_list}{A named list}

\item{datasets}{A Shiny reactive that returns a list of dataset names}
}
\description{
Group Filter Element Server
}
