% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements_ui.R
\name{group_filter_element_ui}
\alias{group_filter_element_ui}
\title{Group Filter Element UI}
\usage{
group_filter_element_ui(id)
}
\arguments{
\item{id}{A shiny ID}
}
\description{
Group Filter Element UI
}
