% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{ici_responder_cohort_obj}
\alias{ici_responder_cohort_obj}
\title{ici_responder_cohort_obj.}
\format{
An object of class \code{Cohort} (inherits from \code{R6}) of length 20.
}
\usage{
ici_responder_cohort_obj
}
\description{
ici_responder_cohort_obj.
}
\keyword{datasets}
