% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_remove_element_server.R
\name{insert_remove_element_server}
\alias{insert_remove_element_server}
\title{Insert Remove Element Server}
\usage{
insert_remove_element_server(
  id,
  element_module,
  element_module_ui,
  remove_ui_event = shiny::reactive(NULL)
)
}
\arguments{
\item{id}{A shiny ID}

\item{element_module}{A shiny server module}

\item{element_module_ui}{A shiny UI module}

\item{remove_ui_event}{A Shiny reactive}
}
\description{
Insert Remove Element Server
}
