% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_remove_element_ui.R
\name{insert_remove_element_ui}
\alias{insert_remove_element_ui}
\title{Insert Remove Element UI}
\usage{
insert_remove_element_ui(id, button_label = "Add UI element")
}
\arguments{
\item{id}{A shiny ID}

\item{button_label}{A string}
}
\description{
Insert Remove Element UI
}
