% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver_associations_multivariate_server.R
\name{multivariate_driver_server}
\alias{multivariate_driver_server}
\title{Multivariate Driver Server}
\usage{
multivariate_driver_server(id, cohort_obj)
}
\arguments{
\item{id}{A string, a module id}

\item{cohort_obj}{A named list}
}
\description{
Multivariate Driver Server
}
