% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver_associations_multivariate_ui.R
\name{multivariate_driver_ui}
\alias{multivariate_driver_ui}
\title{Multivariate Driver UI}
\usage{
multivariate_driver_ui(id)
}
\arguments{
\item{id}{A string, a module id}
}
\description{
Multivariate Driver UI
}
