% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements_server.R
\name{numeric_filter_element_server}
\alias{numeric_filter_element_server}
\title{Numeric Filter Element Server}
\usage{
numeric_filter_element_server(
  id,
  reactive_values,
  module_id,
  numeric_named_list,
  dataset
)
}
\arguments{
\item{id}{A shiny ID}

\item{reactive_values}{a shiny::reactiveValues() object}

\item{module_id}{A unique value for this instance of the element being called}

\item{numeric_named_list}{A named list}

\item{dataset}{A string}
}
\description{
Numeric Filter Element Server
}
