% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements_ui.R
\name{numeric_filter_element_ui}
\alias{numeric_filter_element_ui}
\title{Numeric Filter Element UI}
\usage{
numeric_filter_element_ui(id)
}
\arguments{
\item{id}{A shiny ID}
}
\description{
Numeric Filter Element UI
}
