% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements_server.R
\name{numeric_model_covariate_element_server}
\alias{numeric_model_covariate_element_server}
\title{Numerical Model Covariate Element Server}
\usage{
numeric_model_covariate_element_server(
  id,
  reactive_values,
  module_id,
  covariate_list
)
}
\arguments{
\item{id}{A shiny ID}

\item{reactive_values}{a shiny::reactiveValues() object}

\item{module_id}{A unique value for this instance of the element being called}

\item{covariate_list}{A named list}
}
\description{
Numerical Model Covariate Element Server
}
