% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements_ui.R
\name{numeric_model_covariate_element_ui}
\alias{numeric_model_covariate_element_ui}
\title{Numerical Model Covariate Element UI}
\usage{
numeric_model_covariate_element_ui(id)
}
\arguments{
\item{id}{A shiny ID}
}
\description{
Numerical Model Covariate Element UI
}
