% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{pcawg_immune_subtype_cohort_obj}
\alias{pcawg_immune_subtype_cohort_obj}
\title{pcawg_immune_subtype_cohort_obj.}
\format{
An object of class \code{Cohort} (inherits from \code{R6}) of length 20.
}
\usage{
pcawg_immune_subtype_cohort_obj
}
\description{
pcawg_immune_subtype_cohort_obj.
}
\keyword{datasets}
