% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_histogram.R
\name{plotly_histogram}
\alias{plotly_histogram}
\title{Plotly Histogram}
\usage{
plotly_histogram(
  plot_data,
  x_col = "x",
  xlab = "",
  ylab = "Count",
  title = "",
  source_name = NULL,
  format_func = format_plotly
)
}
\arguments{
\item{plot_data}{A dataframe}

\item{x_col}{A string}

\item{xlab}{A string}

\item{ylab}{A string}

\item{title}{A string}

\item{source_name}{A string or NULL}

\item{format_func}{A function}
}
\description{
Plotly Histogram
}
