% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_object_functions.R
\name{query_feature_values_with_cohort_object}
\alias{query_feature_values_with_cohort_object}
\title{Query Feature Values With Cohort Object}
\usage{
query_feature_values_with_cohort_object(
  cohort_object,
  features = NA,
  feature_classes = NA
)
}
\arguments{
\item{cohort_object}{A named list}

\item{features}{A vector of strings}

\item{feature_classes}{A vector of strings}
}
\description{
Query Feature Values With Cohort Object
}
