% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_object_functions.R
\name{query_gene_expression_with_cohort_object}
\alias{query_gene_expression_with_cohort_object}
\title{Query Gene Expression With Cohort Object}
\usage{
query_gene_expression_with_cohort_object(
  cohort_object,
  gene_types = NA,
  entrez = NA
)
}
\arguments{
\item{cohort_object}{A named list}

\item{gene_types}{A vector of strings}

\item{entrez}{A vector of integers}
}
\description{
Query Gene Expression With Cohort Object
}
