% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tranformations.R
\name{summarise_barplot_se}
\alias{summarise_barplot_se}
\title{Summarise Barplot Standard Error}
\usage{
summarise_barplot_se(data, title)
}
\arguments{
\item{data}{A dataframe with columns "group", "feature", "feature_value"}

\item{title}{A string}
}
\description{
Summarise Barplot Standard Error
}
