% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summarise_tbl_at_column}
\alias{summarise_tbl_at_column}
\title{Summarise Tibble At Column}
\usage{
summarise_tbl_at_column(tbl, column, grouping_columns, function_names)
}
\arguments{
\item{tbl}{A tibble}

\item{column}{A string, a column in the tibble}

\item{grouping_columns}{A vector of strings, columns in the tibble}

\item{function_names}{A function fun, a quosure style lambda ~ fun(.) or a
list of either form.}
}
\description{
Summarise Tibble At Column
}
