% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{tcga_bin_cohort_obj_small}
\alias{tcga_bin_cohort_obj_small}
\title{tcga_bin_cohort_obj_small.}
\format{
An object of class \code{Cohort} (inherits from \code{R6}) of length 20.
}
\usage{
tcga_bin_cohort_obj_small
}
\description{
tcga_bin_cohort_obj_small.
}
\keyword{datasets}
