% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{tcga_tcga_study_cohort_obj}
\alias{tcga_tcga_study_cohort_obj}
\title{tcga_tcga_study_cohort_obj.}
\format{
An object of class \code{Cohort} (inherits from \code{R6}) of length 20.
}
\usage{
tcga_tcga_study_cohort_obj
}
\description{
tcga_tcga_study_cohort_obj.
}
\keyword{datasets}
