% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver_associations_univariate_server.R
\name{univariate_driver_server}
\alias{univariate_driver_server}
\title{Univariate Driver Server}
\usage{
univariate_driver_server(id, cohort_obj)
}
\arguments{
\item{id}{A string, a module id}

\item{cohort_obj}{A named list}
}
\description{
Univariate Driver Server
}
