% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver_associations_univariate_ui.R
\name{univariate_driver_ui}
\alias{univariate_driver_ui}
\title{Univariate Driver UI}
\usage{
univariate_driver_ui(id)
}
\arguments{
\item{id}{A string, a module id}
}
\description{
Univariate Driver UI
}
