% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{upload_cohort_obj}
\alias{upload_cohort_obj}
\title{upload_cohort_obj.}
\format{
An object of class \code{UploadCohort} (inherits from \code{R6}) of length 20.
}
\usage{
upload_cohort_obj
}
\description{
upload_cohort_obj.
}
\keyword{datasets}
