% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot_server.R
\name{barplot_server}
\alias{barplot_server}
\title{Barplot Server}
\usage{
barplot_server(
  id,
  plot_data,
  barplot_xlab = shiny::reactive(""),
  barplot_ylab = shiny::reactive(""),
  barplot_title = shiny::reactive(""),
  barplot_label = shiny::reactive("Feature"),
  drilldown = shiny::reactive(F)
)
}
\arguments{
\item{id}{Module ID}

\item{plot_data}{A shiny::reactive that returns a dataframe with columns
"sample", "group", "feature", "feature_value", and optional columns,
"group_description", and "feature_class"}

\item{barplot_xlab}{A shiny::reactive that returns a string}

\item{barplot_ylab}{A shiny::reactive that returns a string}

\item{barplot_title}{A shiny::reactive that returns a string}

\item{barplot_label}{A shiny::reactive that returns a string}

\item{drilldown}{A shiny::reactive that returns True or False}
}
\description{
Barplot Server
}
