% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drilldown_scatterplot_server.R
\name{drilldown_scatterplot_server}
\alias{drilldown_scatterplot_server}
\title{Drilldown Scatterplot Server}
\usage{
drilldown_scatterplot_server(id, data, eventdata)
}
\arguments{
\item{id}{Module ID}

\item{data}{A shiny::reactive that returns a dataframe with columns
"sample", "x", "y", "color"}

\item{eventdata}{A shiny::reactive that returns a dataframe with column
"x"}
}
\description{
Drilldown Scatterplot Server
}
